worker-service/requirements.txt
redis
psycopg2-binary
pytest
